+++ src/providers/twitch/pubsubmessages/AutoMod.cpp
@@ -2,6 +2,8 @@
 
 #include "util/QMagicEnum.hpp"
 
+#include <QJsonArray>
+
 namespace chatterino {
 
 PubSubAutoModQueueMessage::PubSubAutoModQueueMessage(const QJsonObject &root)
@@ -41,6 +43,27 @@ PubSubAutoModQueueMessage::PubSubAutoModQueueMessage(const QJsonObject &root)
         messageSender.value("display_name").toString();
     this->senderUserChatColor =
         QColor(messageSender.value("chat_color").toString());
+
+    if (this->reason == Reason::BlockedTerm)
+    {
+        // Attempt to read the blocked term(s) that caused this message to be blocked
+        const auto caughtMessageReason =
+            data.value("caught_message_reason").toObject();
+        const auto blockedTermFailure =
+            caughtMessageReason.value("blocked_term_failure").toObject();
+        const auto termsFound =
+            blockedTermFailure.value("terms_found").toArray();
+
+        for (const auto &termValue : termsFound)
+        {
+            const auto term = termValue.toObject();
+            const auto termText = term.value("text").toString();
+            if (!termText.isEmpty())
+            {
+                this->blockedTermsFound.push_back(termText);
+            }
+        }
+    }
 }
 
 }  // namespace chatterino