+++ src/singletons/Settings.hpp
@@ -347,6 +347,10 @@ public:
         "/streamerMode/supressLiveNotifications", false};
     BoolSetting streamerModeSuppressInlineWhispers = {
         "/streamerMode/suppressInlineWhispers", true};
+    BoolSetting streamerModeHideBlockedTermText = {
+        "/streamerMode/hideBlockedTermText",
+        true,
+    };
 
     /// Ignored Phrases
     QStringSetting ignoredPhraseReplace = {"/ignore/ignoredPhraseReplace",